Add parent directory to paths
